Assignermtn 1 fro Programming 2
Employee.h is a header file used to create the sturct employee for the first assignment. PA1 is a program that uses an array to store the names, SSN, department, role, and salary of employees and then display them in a tabular form. the salary, role, and SSN are all randomly generated